:sparkles: Mod: 개인정보 약관 뷰어 페이지 수정
