+++ core/navigation/src/main/java/com/poptato/navigation/NavRoutes.kt
@@ -26,6 +26,7 @@ sealed class NavRoutes(val route: String) {
     data object UserDataScreen: NavRoutes("user_data")
     data object NoticeViewScreen: NavRoutes("notice_viewer")
     data object FAQViewScreen: NavRoutes("faq_viewer")
+    data object PolicyViewScreen: NavRoutes("policy_viewer")
 
     // Today Graph
     data object TodayGraph: NavRoutes("today_graph")
